+++ codex/views/opds/binary.py
@@ -6,10 +6,8 @@ from codex.views.opds.auth import OPDSAuthMixin
 from codex.views.reader.page import ReaderPageView
 
 
-class OPDSCoverView(CoverView):
+class OPDSCoverView(OPDSAuthMixin, CoverView):
     """Cover View with Basic Auth."""
-
-    # Panels sometimes doesn't send auth with cover requests.
 
 
 class OPDSDownloadView(OPDSAuthMixin, DownloadView):