+++ frontend/src/api/v3/browser.js
@@ -1,3 +1,5 @@
+import { cloneDeep } from "lodash";
+
 import { serializeParams } from "@/api/v3/common";
 
 import { HTTP } from "./base";
@@ -28,7 +30,7 @@ export const getCoverSrc = ({ group, pks }, data, ts) => {
 const getMetadata = ({ group, pks }, settings) => {
   const pkList = pks.join(",");
   const mtime = Math.max(group.mtime, settings.mtime);
-  const data = { ...settings };
+  const data = cloneDeep(settings);
   delete data.mtime;
   const params = serializeParams(data, mtime);
   return HTTP.get(`/${group}/${pkList}/metadata`, { params });