+++ frontend/src/components/admin/create-update-dialog/time-text-field.vue
@@ -18,8 +18,7 @@ export default {
       FORMAT: "DDD HH:mm:SS",
       timeRules: [
         (v) =>
-          // eslint-disable-next-line security/detect-unsafe-regex
-          /^(\d{1,3} )?([01]\d|2[0-3]):([0-5]\d):([0-5]\d)$/.test(v) ||
+          /^\d{0,3}?[01]\d|2[0-3](?::[0-5]\d){2}$/.test(v) ||
           `Invalid time format ${this.FORMAT}`,
       ],
     };